requirements.txt
imgrender==0.0.4
matplotlib==3.7.1
mrcnn==0.2
numpy==1.23.5
osgeo==0.0.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
scikit_learn==1.2.2
scikit_image
tensorflow==2.12.0
torchvision==0.15.2
